+++ src/tests/extension/util/helpers/format.test.ts
@@ -1,50 +1,13 @@
 import { describe, expect, it } from "vitest";
 import {
-  timeStrToMS,
-  msToTimeStr,
   formatAmount,
   formatDonation,
   formatPlayers,
   formatTwitch,
   getCurrentGame,
-  selectPrizeFromTier,
-  selectNextBid,
-} from "../../../extension/util/helpers";
+} from "../../../../extension/util/helpers/format";
 
-describe("helpers", () => {
-  describe("timeStrToMS", () => {
-    it("converts HH:MM:SS format to milliseconds", () => {
-      expect(timeStrToMS("00:00:00")).toBe(0);
-      expect(timeStrToMS("00:00:01")).toBe(1_000);
-      expect(timeStrToMS("00:01:00")).toBe(60_000);
-      expect(timeStrToMS("01:00:00")).toBe(3_600_000);
-      expect(timeStrToMS("01:02:03")).toBe(
-        1 * 3_600_000 + 2 * 60_000 + 3 * 1_000
-      );
-    });
-
-    it("treats MM:SS format as 0 hours", () => {
-      expect(timeStrToMS("00:30")).toBe(30_000);
-      expect(timeStrToMS("01:30")).toBe(90_000);
-    });
-  });
-
-  describe("msToTimeStr", () => {
-    it("formats times less than an hour as MM:SS", () => {
-      expect(msToTimeStr(0)).toBe("00:00");
-      expect(msToTimeStr(1_000)).toBe("00:01");
-      expect(msToTimeStr(61_000)).toBe("01:01");
-      expect(msToTimeStr(59_999)).toBe("00:59");
-    });
-
-    it("formats times of an hour or more as H:MM:SS", () => {
-      expect(msToTimeStr(3_600_000)).toBe("1:00:00");
-      expect(msToTimeStr(3_660_000)).toBe("1:01:00");
-      expect(msToTimeStr(3_661_000)).toBe("1:01:01");
-      expect(msToTimeStr(10 * 3_600_000 + 5 * 60_000 + 7_000)).toBe("10:05:07");
-    });
-  });
-
+describe("format helpers", () => {
   describe("formatAmount", () => {
     it("formats integer amounts without fractional digits", () => {
       expect(formatAmount(0)).toBe("0 PLN");
@@ -243,126 +206,5 @@ describe("helpers", () => {
       expect(getCurrentGame({})).toBe("Brak ustawionej gry");
     });
   });
-
-  describe("selectPrizeFromTier", () => {
-    const prizes = [
-      { id: 1, name: "Prize A", minimumBid: 10 },
-      { id: 2, name: "Prize B", minimumBid: 10 },
-      { id: 3, name: "Prize C", minimumBid: 25 },
-      { id: 4, name: "Prize D", minimumBid: 50 },
-      { id: 5, name: "Prize E", minimumBid: 10 },
-    ];
-
-    it("returns undefined for empty prizes array", () => {
-      const result = selectPrizeFromTier([], 10, 0);
-      expect(result.prize).toBeUndefined();
-      expect(result.newIndex).toBe(0);
-    });
-
-    it("returns undefined when no prizes match the tier", () => {
-      const result = selectPrizeFromTier(prizes, 100, 0);
-      expect(result.prize).toBeUndefined();
-      expect(result.newIndex).toBe(0);
-    });
-
-    it("returns the first prize at index 0", () => {
-      const result = selectPrizeFromTier(prizes, 10, 0);
-      expect(result.prize).toEqual({ id: 1, name: "Prize A", minimumBid: 10 });
-      expect(result.newIndex).toBe(0);
-    });
-
-    it("returns the correct prize at a given index", () => {
-      const result = selectPrizeFromTier(prizes, 10, 1);
-      expect(result.prize).toEqual({ id: 2, name: "Prize B", minimumBid: 10 });
-      expect(result.newIndex).toBe(1);
-    });
-
-    it("wraps index when it exceeds available prizes in tier", () => {
-      // There are 3 prizes in tier 10 (indices 0, 1, 2)
-      const result = selectPrizeFromTier(prizes, 10, 3);
-      expect(result.prize).toEqual({ id: 1, name: "Prize A", minimumBid: 10 });
-      expect(result.newIndex).toBe(0);
-    });
-
-    it("wraps index when it far exceeds available prizes", () => {
-      const result = selectPrizeFromTier(prizes, 10, 100);
-      expect(result.prize).toEqual({ id: 1, name: "Prize A", minimumBid: 10 });
-      expect(result.newIndex).toBe(0);
-    });
-
-    it("returns the only prize in a single-prize tier", () => {
-      const result = selectPrizeFromTier(prizes, 50, 0);
-      expect(result.prize).toEqual({ id: 4, name: "Prize D", minimumBid: 50 });
-      expect(result.newIndex).toBe(0);
-    });
-
-    it("wraps immediately for single-prize tier at index 1", () => {
-      const result = selectPrizeFromTier(prizes, 50, 1);
-      expect(result.prize).toEqual({ id: 4, name: "Prize D", minimumBid: 50 });
-      expect(result.newIndex).toBe(0);
-    });
-  });
-
-  describe("selectNextBid", () => {
-    const bids = [
-      { id: 1, name: "Bid A" },
-      { id: 2, name: "Bid B" },
-      { id: 3, name: "Bid C" },
-    ];
-
-    it("returns showPanel false for empty bids array", () => {
-      const result = selectNextBid([], 0, false);
-      expect(result.bid).toBeUndefined();
-      expect(result.newIndex).toBe(0);
-      expect(result.showPanel).toBe(false);
-    });
-
-    it("returns showPanel false for empty bids even if panel was shown", () => {
-      const result = selectNextBid([], 0, true);
-      expect(result.bid).toBeUndefined();
-      expect(result.showPanel).toBe(false);
-    });
-
-    it("returns first bid when panel is not currently shown", () => {
-      const result = selectNextBid(bids, 2, false);
-      expect(result.bid).toEqual({ id: 1, name: "Bid A" });
-      expect(result.newIndex).toBe(0);
-      expect(result.showPanel).toBe(true);
-    });
-
-    it("advances to next bid when panel is already shown", () => {
-      const result = selectNextBid(bids, 0, true);
-      expect(result.bid).toEqual({ id: 2, name: "Bid B" });
-      expect(result.newIndex).toBe(1);
-      expect(result.showPanel).toBe(true);
-    });
-
-    it("continues advancing through bids", () => {
-      const result = selectNextBid(bids, 1, true);
-      expect(result.bid).toEqual({ id: 3, name: "Bid C" });
-      expect(result.newIndex).toBe(2);
-      expect(result.showPanel).toBe(true);
-    });
-
-    it("wraps to first bid after reaching the end", () => {
-      const result = selectNextBid(bids, 2, true);
-      expect(result.bid).toEqual({ id: 1, name: "Bid A" });
-      expect(result.newIndex).toBe(0);
-      expect(result.showPanel).toBe(true);
-    });
-
-    it("handles single bid array", () => {
-      const singleBid = [{ id: 1, name: "Only Bid" }];
-      
-      // First call (panel not shown)
-      const result1 = selectNextBid(singleBid, 0, false);
-      expect(result1.bid).toEqual({ id: 1, name: "Only Bid" });
-      expect(result1.newIndex).toBe(0);
-      
-      // Second call (panel shown, should wrap to same bid)
-      const result2 = selectNextBid(singleBid, 0, true);
-      expect(result2.bid).toEqual({ id: 1, name: "Only Bid" });
-      expect(result2.newIndex).toBe(0);
-    });
-  });
 });
+
