+++ src/tests/extension/util/helpers/time.test.ts
@@ -1,0 +1,41 @@
+import { describe, expect, it } from "vitest";
+import {
+  timeStrToMS,
+  msToTimeStr,
+} from "../../../../extension/util/helpers/time";
+
+describe("time helpers", () => {
+  describe("timeStrToMS", () => {
+    it("converts HH:MM:SS format to milliseconds", () => {
+      expect(timeStrToMS("00:00:00")).toBe(0);
+      expect(timeStrToMS("00:00:01")).toBe(1_000);
+      expect(timeStrToMS("00:01:00")).toBe(60_000);
+      expect(timeStrToMS("01:00:00")).toBe(3_600_000);
+      expect(timeStrToMS("01:02:03")).toBe(
+        1 * 3_600_000 + 2 * 60_000 + 3 * 1_000
+      );
+    });
+
+    it("treats MM:SS format as 0 hours", () => {
+      expect(timeStrToMS("00:30")).toBe(30_000);
+      expect(timeStrToMS("01:30")).toBe(90_000);
+    });
+  });
+
+  describe("msToTimeStr", () => {
+    it("formats times less than an hour as MM:SS", () => {
+      expect(msToTimeStr(0)).toBe("00:00");
+      expect(msToTimeStr(1_000)).toBe("00:01");
+      expect(msToTimeStr(61_000)).toBe("01:01");
+      expect(msToTimeStr(59_999)).toBe("00:59");
+    });
+
+    it("formats times of an hour or more as H:MM:SS", () => {
+      expect(msToTimeStr(3_600_000)).toBe("1:00:00");
+      expect(msToTimeStr(3_660_000)).toBe("1:01:00");
+      expect(msToTimeStr(3_661_000)).toBe("1:01:01");
+      expect(msToTimeStr(10 * 3_600_000 + 5 * 60_000 + 7_000)).toBe("10:05:07");
+    });
+  });
+});
+
